src/main/java/com/proyecto/entity/Usuario.java
package com.proyecto.entity;

import jakarta.persistence.*;

@Entity
@Table(name = "USUARIO")
public class Usuario {
  @Id
  private Integer id;

  private String correo;

  private String contrasena;

  @OneToOne(mappedBy = "usuario", cascade = CascadeType.ALL)
  private Empleado empleado;

  public Integer getId() {
    return id;
  }

  public void setId(Integer id) {
    this.id = id;
  }

  public String getCorreo() {
    return correo;
  }

  public void setCorreo(String correo) {
    this.correo = correo;
  }

  public String getContrasena() {
    return contrasena;
  }

  public void setContrasena(String contrasena) {
    this.contrasena = contrasena;
  }

  public Empleado getEmpleado() {
    return empleado;
  }

  public void setEmpleado(Empleado empleado) {
    this.empleado = empleado;
  }
}
